requirements.txt
google-generativeai
python-dotenv
pillow
fastapi
uvicorn
python-multipart
huggingface_hub
supabase
flask
flask-cors